// src/adapters/github/generated/github-types.ts
export type ProjectV2ItemType =
  | "ISSUE"
  | "PULL_REQUEST"
  | "DRAFT_ISSUE"
  | "REDACTED";
// ...
export type ProjectV2FieldConfiguration =
  | ProjectV2Field
  | ProjectV2IterationField
  | ProjectV2SingleSelectField;
export type ProjectV2IssueFieldValues =
  | IssueFieldDateValue
  | IssueFieldNumberValue
  | IssueFieldSingleSelectValue
  | IssueFieldTextValue;
export type ProjectV2ItemContent = DraftIssue | Issue | PullRequest;
export type ProjectV2ItemFieldValue =
  | ProjectV2ItemFieldDateValue
  | ProjectV2ItemFieldIterationValue
  | ProjectV2ItemFieldLabelValue
  | ProjectV2ItemFieldMilestoneValue
  | ProjectV2ItemFieldNumberValue
  | ProjectV2ItemFieldPullRequestValue
  | ProjectV2ItemFieldRepositoryValue
  | ProjectV2ItemFieldReviewerValue
  | ProjectV2ItemFieldSingleSelectValue
  | ProjectV2ItemFieldTextValue
  | ProjectV2ItemFieldUserValue
  | ProjectV2ItemIssueFieldValue;
// ...
export interface Actor {
  avatarUrl: string;
  login: string;
  resourcePath: string;
  url: string;
}
// ...
export interface DraftIssue {
  assignees?: UserConnection;
  body?: string;
  bodyHTML?: string;
  bodyText?: string;
  createdAt?: string;
  creator?: Actor | null;
  id?: string;
  projectV2Items?: ProjectV2ItemConnection;
  projectsV2?: ProjectV2Connection;
  title?: string;
  updatedAt?: string;
}
// ...
export interface Issue {
  activeLockReason?: LockReason | null;
  assignedActors?: AssigneeConnection;
  assignees?: UserConnection;
  author?: Actor | null;
  authorAssociation?: CommentAuthorAssociation;
  blockedBy?: IssueConnection;
  blocking?: IssueConnection;
  body?: string;
  bodyHTML?: string;
  bodyResourcePath?: string;
  bodyText?: string;
  bodyUrl?: string;
  closed?: boolean;
  closedAt?: string | null;
  closedByPullRequestsReferences?: PullRequestConnection | null;
  comments?: IssueCommentConnection;
  createdAt?: string;
  createdViaEmail?: boolean;
  databaseId?: number | null;
  duplicateOf?: Issue | null;
  editor?: Actor | null;
  fullDatabaseId?: string | null;
  hovercard?: Hovercard;
  id?: string;
  includesCreatedEdit?: boolean;
  isPinned?: boolean | null;
  isReadByViewer?: boolean | null;
  issueDependenciesSummary?: IssueDependenciesSummary;
  issueFieldValues?: IssueFieldValueConnection | null;
  issueType?: IssueType | null;
  labels?: LabelConnection | null;
  lastEditedAt?: string | null;
  linkedBranches?: LinkedBranchConnection;
  locked?: boolean;
  milestone?: Milestone | null;
  number?: number;
  parent?: Issue | null;
  participants?: UserConnection;
  pinnedIssueComment?: PinnedIssueComment | null;
  projectCards?: ProjectCardConnection;
  projectItems?: ProjectV2ItemConnection | null;
  projectV2?: ProjectV2 | null;
  projectsV2?: ProjectV2Connection;
  publishedAt?: string | null;
  reactionGroups?: Array<ReactionGroup> | null;
  reactions?: ReactionConnection;
  repository?: Repository;
  resourcePath?: string;
  state?: IssueState;
  stateReason?: IssueStateReason | null;
  subIssues?: IssueConnection;
  subIssuesSummary?: SubIssuesSummary;
  suggestedActors?: AssigneeConnection;
  timeline?: IssueTimelineConnection;
  timelineItems?: IssueTimelineItemsConnection;
  title?: string;
  titleHTML?: string;
  trackedInIssues?: IssueConnection;
  trackedIssues?: IssueConnection;
  trackedIssuesCount?: number;
  updatedAt?: string;
  url?: string;
  userContentEdits?: UserContentEditConnection | null;
  viewerCanClose?: boolean;
  viewerCanDelete?: boolean;
  viewerCanLabel?: boolean;
  viewerCanReact?: boolean;
  viewerCanReopen?: boolean;
  viewerCanSetFields?: boolean | null;
  viewerCanSubscribe?: boolean;
  viewerCanUpdate?: boolean;
  viewerCannotUpdateReasons?: Array<CommentCannotUpdateReason>;
  viewerDidAuthor?: boolean;
  viewerSubscription?: SubscriptionState | null;
  viewerThreadSubscriptionFormAction?: ThreadSubscriptionFormAction | null;
  viewerThreadSubscriptionStatus?: ThreadSubscriptionState | null;
}
// ...
export interface PageInfo {
  endCursor?: string | null;
  hasNextPage?: boolean;
  hasPreviousPage?: boolean;
  startCursor?: string | null;
}
// ...
export interface ProjectV2 {
  closed?: boolean;
  closedAt?: string | null;
  createdAt?: string;
  creator?: Actor | null;
  databaseId?: number | null;
  field?: ProjectV2FieldConfiguration | null;
  fields?: ProjectV2FieldConfigurationConnection;
  fullDatabaseId?: string | null;
  id?: string;
  items?: ProjectV2ItemConnection;
  number?: number;
  owner?: ProjectV2Owner;
  public?: boolean;
  readme?: string | null;
  repositories?: RepositoryConnection;
  resourcePath?: string;
  shortDescription?: string | null;
  statusUpdates?: ProjectV2StatusUpdateConnection;
  teams?: TeamConnection;
  template?: boolean;
  title?: string;
  updatedAt?: string;
  url?: string;
  view?: ProjectV2View | null;
  viewerCanClose?: boolean;
  viewerCanReopen?: boolean;
  viewerCanUpdate?: boolean;
  views?: ProjectV2ViewConnection;
  workflow?: ProjectV2Workflow | null;
  workflows?: ProjectV2WorkflowConnection;
}
// ...
export interface ProjectV2FieldConfigurationConnection {
  edges?: Array<ProjectV2FieldConfigurationEdge | null> | null;
  nodes?: Array<ProjectV2FieldConfiguration | null> | null;
  pageInfo?: PageInfo;
  totalCount?: number;
}
// ...
export interface ProjectV2Item {
  content?: ProjectV2ItemContent | null;
  createdAt?: string;
  creator?: Actor | null;
  databaseId?: number | null;
  fieldValueByName?: ProjectV2ItemFieldValue | null;
  fieldValues?: ProjectV2ItemFieldValueConnection;
  fullDatabaseId?: string | null;
  id?: string;
  isArchived?: boolean;
  project?: ProjectV2;
  type?: ProjectV2ItemType;
  updatedAt?: string;
}

export interface ProjectV2ItemConnection {
  edges?: Array<ProjectV2ItemEdge | null> | null;
  nodes?: Array<ProjectV2Item | null> | null;
  pageInfo?: PageInfo;
  totalCount?: number;
}

export interface ProjectV2ItemEdge {
  cursor?: string;
  node?: ProjectV2Item | null;
}

export interface ProjectV2ItemFieldDateValue {
  createdAt?: string;
  creator?: Actor | null;
  databaseId?: number | null;
  date?: string | null;
  field?: ProjectV2FieldConfiguration;
  id?: string;
  item?: ProjectV2Item;
  updatedAt?: string;
}

export interface ProjectV2ItemFieldIterationValue {
  createdAt?: string;
  creator?: Actor | null;
  databaseId?: number | null;
  duration?: number;
  field?: ProjectV2FieldConfiguration;
  id?: string;
  item?: ProjectV2Item;
  iterationId?: string;
  startDate?: string;
  title?: string;
  titleHTML?: string;
  updatedAt?: string;
}

export interface ProjectV2ItemFieldLabelValue {
  field?: ProjectV2FieldConfiguration;
  labels?: LabelConnection | null;
}

export interface ProjectV2ItemFieldMilestoneValue {
  field?: ProjectV2FieldConfiguration;
  milestone?: Milestone | null;
}

export interface ProjectV2ItemFieldNumberValue {
  createdAt?: string;
  creator?: Actor | null;
  databaseId?: number | null;
  field?: ProjectV2FieldConfiguration;
  id?: string;
  item?: ProjectV2Item;
  number?: number | null;
  updatedAt?: string;
}

export interface ProjectV2ItemFieldPullRequestValue {
  field?: ProjectV2FieldConfiguration;
  pullRequests?: PullRequestConnection | null;
}

export interface ProjectV2ItemFieldRepositoryValue {
  field?: ProjectV2FieldConfiguration;
  repository?: Repository | null;
}

export interface ProjectV2ItemFieldReviewerValue {
  field?: ProjectV2FieldConfiguration;
  reviewers?: RequestedReviewerConnection | null;
}

export interface ProjectV2ItemFieldSingleSelectValue {
  color?: ProjectV2SingleSelectFieldOptionColor;
  createdAt?: string;
  creator?: Actor | null;
  databaseId?: number | null;
  description?: string | null;
  descriptionHTML?: string | null;
  field?: ProjectV2FieldConfiguration;
  id?: string;
  item?: ProjectV2Item;
  name?: string | null;
  nameHTML?: string | null;
  optionId?: string | null;
  updatedAt?: string;
}

export interface ProjectV2ItemFieldTextValue {
  createdAt?: string;
  creator?: Actor | null;
  databaseId?: number | null;
  field?: ProjectV2FieldConfiguration;
  id?: string;
  item?: ProjectV2Item;
  text?: string | null;
  updatedAt?: string;
}

export interface ProjectV2ItemFieldUserValue {
  field?: ProjectV2FieldConfiguration;
  users?: UserConnection | null;
}

export interface ProjectV2ItemFieldValueConnection {
  edges?: Array<ProjectV2ItemFieldValueEdge | null> | null;
  nodes?: Array<ProjectV2ItemFieldValue | null> | null;
  pageInfo?: PageInfo;
  totalCount?: number;
}

export interface ProjectV2ItemFieldValueEdge {
  cursor?: string;
  node?: ProjectV2ItemFieldValue | null;
}

export interface ProjectV2ItemIssueFieldValue {
  field?: ProjectV2FieldConfiguration;
  issueFieldValue?: ProjectV2IssueFieldValues | null;
}
// ...
export interface PullRequest {
  activeLockReason?: LockReason | null;
  additions?: number;
  assignedActors?: AssigneeConnection;
  assignees?: UserConnection;
  author?: Actor | null;
  authorAssociation?: CommentAuthorAssociation;
  autoMergeRequest?: AutoMergeRequest | null;
  baseRef?: Ref | null;
  baseRefName?: string;
  baseRefOid?: string;
  baseRepository?: Repository | null;
  body?: string;
  bodyHTML?: string;
  bodyText?: string;
  canBeRebased?: boolean;
  changedFiles?: number;
  checksResourcePath?: string;
  checksUrl?: string;
  closed?: boolean;
  closedAt?: string | null;
  closingIssuesReferences?: IssueConnection | null;
  comments?: IssueCommentConnection;
  commits?: PullRequestCommitConnection;
  createdAt?: string;
  createdViaEmail?: boolean;
  databaseId?: number | null;
  deletions?: number;
  editor?: Actor | null;
  files?: PullRequestChangedFileConnection | null;
  fullDatabaseId?: string | null;
  headRef?: Ref | null;
  headRefName?: string;
  headRefOid?: string;
  headRepository?: Repository | null;
  headRepositoryOwner?: RepositoryOwner | null;
  hovercard?: Hovercard;
  id?: string;
  includesCreatedEdit?: boolean;
  isCrossRepository?: boolean;
  isDraft?: boolean;
  isInMergeQueue?: boolean;
  isMergeQueueEnabled?: boolean;
  isReadByViewer?: boolean | null;
  labels?: LabelConnection | null;
  lastEditedAt?: string | null;
  latestOpinionatedReviews?: PullRequestReviewConnection | null;
  latestReviews?: PullRequestReviewConnection | null;
  locked?: boolean;
  maintainerCanModify?: boolean;
  mergeCommit?: Commit | null;
  mergeQueue?: MergeQueue | null;
  mergeQueueEntry?: MergeQueueEntry | null;
  mergeStateStatus?: MergeStateStatus;
  mergeable?: MergeableState;
  merged?: boolean;
  mergedAt?: string | null;
  mergedBy?: Actor | null;
  milestone?: Milestone | null;
  number?: number;
  participants?: UserConnection;
  permalink?: string;
  potentialMergeCommit?: Commit | null;
  projectCards?: ProjectCardConnection;
  projectItems?: ProjectV2ItemConnection | null;
  projectV2?: ProjectV2 | null;
  projectsV2?: ProjectV2Connection;
  publishedAt?: string | null;
  reactionGroups?: Array<ReactionGroup> | null;
  reactions?: ReactionConnection;
  repository?: Repository;
  resourcePath?: string;
  revertResourcePath?: string;
  revertUrl?: string;
  reviewDecision?: PullRequestReviewDecision | null;
  reviewRequests?: ReviewRequestConnection | null;
  reviewThreads?: PullRequestReviewThreadConnection;
  reviews?: PullRequestReviewConnection | null;
  state?: PullRequestState;
  statusCheckRollup?: StatusCheckRollup | null;
  suggestedActors?: AssigneeConnection;
  suggestedReviewerActors?: SuggestedReviewerActorConnection;
  suggestedReviewers?: Array<SuggestedReviewer | null>;
  timeline?: PullRequestTimelineConnection;
  timelineItems?: PullRequestTimelineItemsConnection;
  title?: string;
  titleHTML?: string;
  totalCommentsCount?: number | null;
  updatedAt?: string;
  url?: string;
  userContentEdits?: UserContentEditConnection | null;
  viewerCanApplySuggestion?: boolean;
  viewerCanClose?: boolean;
  viewerCanDeleteHeadRef?: boolean;
  viewerCanDisableAutoMerge?: boolean;
  viewerCanEditFiles?: boolean;
  viewerCanEnableAutoMerge?: boolean;
  viewerCanLabel?: boolean;
  viewerCanMergeAsAdmin?: boolean;
  viewerCanReact?: boolean;
  viewerCanReopen?: boolean;
  viewerCanSubscribe?: boolean;
  viewerCanUpdate?: boolean;
  viewerCanUpdateBranch?: boolean;
  viewerCannotUpdateReasons?: Array<CommentCannotUpdateReason>;
  viewerDidAuthor?: boolean;
  viewerLatestReview?: PullRequestReview | null;
  viewerLatestReviewRequest?: ReviewRequest | null;
  viewerMergeBodyText?: string;
  viewerMergeHeadlineText?: string;
  viewerSubscription?: SubscriptionState | null;
}